Feat - In progress Tactic Dashboard
