--- src/stores/filter.js
@@ -1,12 +1,0 @@
-import {defineStore} from 'pinia'
-
-export const useFilterStore = defineStore('filter', {
-    state() {
-        return {
-            dateFilter: null
-        }
-    },
-    actions: {
-
-    }
-})
